+++ lib/src/messages/messages.barrel.dart
@@ -1,0 +1,1 @@
+export 'conflicting_input_border_parts_message.dart';