+++ lib/src/partials/partials.barrel.dart
@@ -1,5 +1,7 @@
 export 'partial_border.dart';
 export 'partial_border_side.dart';
 export 'partial_box_shadow.dart';
-export 'partial_input_border.dart';
+export 'partial_outline_input_border.dart';
+export 'partial_underline_input_border.dart';
+export 'partial_no_input_border.dart';
 export 'partial_shadow.dart';